requirements.txt
faiss-cpu
torch
sentence-transformers
numpy
supabase
python-dotenv